An example of how to generate uniform bath energy spacing is provided in: `input_settings/rspt2spectra_parameters.py`.
A bit of a cleanup, even though more things can easily be improved.

The script `scripts/finiteH0.py` is still not tested in the presence of a self-energy.
One might would like to have an option to use RSPT's local Hamiltonian values in the generated Hamiltonian, instead of the adjusted on-site energies.