The script finiteH0.py can handle general cases of systems also with off-diagonal hybridization functions. The script finiteH0_diag.py assumes the studied system do not have off-diagonal onsite energies (and off-diagonal hybridization functions). The two scripts use different input data (see in the input_settings folder).
The only good reasons to use the finiteH0_diag.py script is that is uses a better algorithm for determining the strength of the hopping parameters.
This means that one can get a reasonable hybridization discretization with a few bath states, if using the finiteH0_diag.py script.
But if one uses the finiteH0.py script, more bath states are at the moment needed, but the algorithm here can easily be improved.

Another limitation at the moment is that spherical harmonics functions are assumed as the impurity orbitals.
This limitation will hopefully be removed soon.